Refactor entity tile components
