+++ context/app/static/js/components/entity-tile/EntityTileBody/index.js
@@ -1,0 +1,3 @@
+import EntityTileBody from './EntityTileBody';
+
+export default EntityTileBody;
