--- context/app/static/js/components/entity-tile/EntityTileBottom/index.js
@@ -1,3 +1,0 @@
-import EntityTileBottom from './EntityTileBottom';
-
-export default EntityTileBottom;
