+++ context/app/static/js/components/entity-tile/EntityTileFooter/index.js
@@ -1,0 +1,3 @@
+import EntityTileFooter from './EntityTileFooter';
+
+export default EntityTileFooter;
