--- context/app/static/js/components/entity-tile/EntityTileTop/index.js
@@ -1,3 +1,0 @@
-import EntityTileTop from './EntityTileTop';
-
-export default EntityTileTop;
