--- context/app/static/js/components/entity-tile/EntityTileTopText/index.js
@@ -1,3 +1,0 @@
-import EntityTileTopText from './EntityTileTopText';
-
-export default EntityTileTopText;
